Move 4 playlist template-var normalizers into modules/output_playlists.py
Fourth refactor pass at modules/output.py.  Rather than a new module,
these four helpers logically belong with the playlist file-entry and
library-selection helpers extracted in PR #1446 -- they finish the
playlist-shaped concerns.  Moved into an existing module rather than
inventing a new one to preserve cohesion.

Functions moved (from output.py to output_playlists.py):
  * _normalize_playlist_template_var_value
  * _normalize_playlist_keyed_template_var_value
  * _collect_playlist_template_variables_from_libraries_data
  * _collect_playlist_file_entries_from_libraries_data

Constants moved:
  * PLAYLIST_SHARED_TEMPLATE_VAR_SPECS
  * PLAYLIST_KEYED_TEMPLATE_VAR_SPECS

These two module-level dicts declare the expected 'value kind' for each
playlist template variable (string / string_list / boolean / integer /
ignore_ids) and drive the normalizers.  They're only used inside the
normalizers, so keeping them together makes maintenance easier.

Dependencies: modules.output_values (from PR #1445) for value coercion.

Public surface: output.py re-imports all six names with '# noqa: F401'.

Impact:
- modules/output.py: 3343 -> 3207 (-136 / -4.1%)
- modules/output_playlists.py: 172 -> 338 (+166 lines)

Cumulative on output.py: 3833 -> 3207 (-626 / -16.3%)

All 1077 tests pass. Ruff + black clean.
